+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v0.8.0
 Library:
 1. [feat] 新增多点触控接口，目前仅 nemu_ipc 实现了多点触控。
+2. [feat] 优化库 import 速度，从 2.9 秒降至 0.92 秒，提升了 68%。
 
 ## v0.7.0
 Library: